fixed definition of fibonacci_sum fsum func

+++ task2/project/fibonacci.h
@@ -1,5 +1,5 @@
 #include <cmath>
 #include <cstdint>
 
-uint64_t fibonacci_sum(unsigned int n);
+uint64_t fibonacci_sum(uint64_t n);
 uint64_t fibonacci(uint64_t pre_last, uint64_t last, uint64_t n);
